chore(jans-cedarling): refactor move parsing `PolicySet` closer to the structure
